requirements.txt
Django==4.1
django-apscheduler==0.6.2
mysqlclient==2.1.1
python-dotenv==1.0.0
pandas==2.0.1
pymysql==1.0.3
snownlp==0.12.3
django-filter==2.3.0
djangorestframework==3.11.1
djangorestframework-simplejwt==4.6.0
drf-yasg==1.21.5